+++ kaithem/src/messagebus.py
@@ -1,11 +1,11 @@
 # SPDX-FileCopyrightText: Copyright 2013 Daniel Dunn
 # SPDX-License-Identifier: GPL-3.0-only
 
-import logging
 import traceback
 import weakref
 
 import cherrypy
+import structlog
 from scullery import messagebus
 
 post_message = messagebus.post_message
@@ -15,6 +15,7 @@ unsubscribe = messagebus.unsubscribe
 normalize_topic = messagebus.normalize_topic
 log = messagebus.log
 MessageBus = messagebus.MessageBus
+logger = structlog.get_logger("system.msgbus")
 
 
 def handleMsgbusError(f, topic, message):
@@ -33,7 +34,7 @@ def handleMsgbusError(f, topic, message):
                 x = hasattr(f, "_kaithemAlreadyPostedNotificatonError")
                 f._kaithemAlreadyPostedNotificatonError = True
                 if not x:
-                    logging.exception("Message subscriber error for " + str(topic))
+                    logger.exception("Message subscriber error for " + str(topic))
             except Exception:
                 print(traceback.format_exc())