+++ kaithem/src/plugin_system.py
@@ -101,7 +101,7 @@ def load_user_plugins():
                         assert spec
                         import_in_thread(spec)
                     except Exception:
-                        logging.exception("Error in user plugin")
+                        logger.exception("Error in user plugin")
 
         for i in range(240000):
             time.sleep(0.001)